apicrudpro/requirements.txt
Django==2.0
django-extensions==2.1.3
django-filter==2.1.0
djangorestframework==3.9.2